fuzz: pull FuzzPk out of miniscript_satisfy target
This should be in a library so it can be used by multiple fuzztests.
